requirements.txt
# Core App Framework
streamlit==1.29.0

# LangChain Core
langchain==0.1.9
langchain-community==0.0.20
langchain-openai==0.0.8
langchain-experimental==0.0.49

# Vector Database
chromadb==0.4.22

# OpenAI API
openai==1.9.0

# Data Handling
pandas==2.0.3
numpy==1.24.3

# For plotting & visualization
matplotlib==3.7.1
plotly==5.15.0

# For loading environment variables
python-dotenv==1.0.0

# For the code execution safety (will be critical later)
python-dotenv==1.0.0